requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
html5lib==1.1
idna==3.6
lxml==4.9.4
numpy==1.26.2
pandas==2.1.4
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tabulate==0.9.0
tzdata==2023.3
urllib3==2.1.0
webencodings==0.5.1
